// src/app/models/cell.ts
export interface CellModel {
    id: number;
    name: string;
    temp: string;
    storeId: number;
    foodId: number;
    shelfLife: string;
    price: number;
    foodName: string;
    foodDescription: string;
    status: string;
    number: number;
    weight: number;
    providerId: number;
    providerName: string;
    providerPrice: number;
    providerUserId: number;
    providerLocation: string;
    isPacked: boolean;
}